Added role checks in Task controller

+++ Back end/GroupProject1/src/main/java/Dreamteam/GroupProject1/controllers/TaskController.java
@@ -1,8 +1,13 @@
 package Dreamteam.GroupTask1.controllers;
 
+import Dreamteam.GroupProject1.controllers.CommentController;
+import Dreamteam.GroupProject1.controllers.Exceptions.UnauthorizedException;
 import Dreamteam.GroupProject1.dto.task.*;
 import Dreamteam.GroupProject1.dto.task.TaskCreateDTO;
 import Dreamteam.GroupProject1.dto.task.TaskDTO;
+import Dreamteam.GroupProject1.models.AppUser;
+import Dreamteam.GroupProject1.models.enums.Role;
+import Dreamteam.GroupProject1.service.AppUserService;
 import Dreamteam.GroupProject1.service.TaskService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +22,7 @@ import java.util.List;
 @RequestMapping("tasks")
 public class TaskController {
     final private TaskService taskService;
+    final private AppUserService userService;
 
     @Autowired
     public TaskController(TaskService taskService) {
@@ -32,7 +38,13 @@ public class TaskController {
 //    }
 
     @PostMapping
-    public ResponseEntity<TaskDTO> createTask(@Valid @RequestBody TaskCreateDTO createDto) {
+    public ResponseEntity<TaskDTO> createTask(@Valid @RequestBody TaskCreateDTO createDto, @RequestParam Long userId) {
+
+        AppUser appUser = userService.getUserById(userId);
+
+        if (appUser.hasRole(Role.CLIENT))
+            throw new UnauthorizedException("Only Project Managers and Developers can create comments.");
+
         TaskDTO created = taskService.createTask(createDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
     }
@@ -46,7 +58,14 @@ public class TaskController {
     @PatchMapping("/{id}")
     public ResponseEntity<TaskDTO> updateTask(
             @PathVariable Long id,
-            @Valid @RequestBody TaskUpdateDTO updateDto) {
+            @Valid @RequestBody TaskUpdateDTO updateDto,
+            @RequestParam Long userId) {
+
+        AppUser appUser = userService.getUserById(userId);
+
+        if (appUser.hasRole(Role.CLIENT))
+            throw new UnauthorizedException("Only Project Managers and Developers can create comments.");
+
         TaskDTO updated = taskService.updateTask(id, updateDto);
         return ResponseEntity.ok(updated);
     }
@@ -55,6 +74,12 @@ public class TaskController {
     public ResponseEntity<TaskDTO> addWatchingUser(
             @PathVariable Long taskId,
             @PathVariable Long userId) {
+
+        AppUser appUser = userService.getUserById(userId);
+
+        if (appUser.hasRole(Role.CLIENT))
+            throw new UnauthorizedException("Only Project Managers and Developers can create comments.");
+
         TaskDTO updated = taskService.addWatchingUser(taskId, userId);
         return ResponseEntity.ok(updated);
     }
@@ -63,12 +88,25 @@ public class TaskController {
     public ResponseEntity<TaskDTO> removeWatchingUser(
             @PathVariable Long taskId,
             @PathVariable Long userId) {
+
+        AppUser appUser = userService.getUserById(userId);
+
+        if (appUser.hasRole(Role.CLIENT))
+            throw new UnauthorizedException("Only Project Managers and Developers can create comments.");
+
         TaskDTO updated = taskService.removeWatchingUser(taskId, userId);
         return ResponseEntity.ok(updated);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<TaskDTO> deleteTask(@PathVariable Long id) {
+    public ResponseEntity<TaskDTO> deleteTask(@PathVariable Long id,
+                                              @RequestParam Long userId) {
+
+        AppUser appUser = userService.getUserById(userId);
+
+        if (appUser.hasRole(Role.CLIENT))
+            throw new UnauthorizedException("Only Project Managers and Developers can create comments.");
+
         TaskDTO task = taskService.findById(id);
         taskService.deleteTask(id);
         return ResponseEntity.ok(task);
